Issue-9660 - Extract array_to_string and string_to_array from kernels and udf containers
