+++ datafusion/functions-array/src/lib.rs
@@ -39,6 +39,7 @@ mod remove;
 mod replace;
 mod rewrite;
 mod set_ops;
+mod string;
 mod udf;
 mod utils;
 
@@ -73,6 +74,8 @@ pub mod expr_fn {
     pub use super::set_ops::array_distinct;
     pub use super::set_ops::array_intersect;
     pub use super::set_ops::array_union;
+    pub use super::string::array_to_string;
+    pub use super::string::string_to_array;
     pub use super::udf::array_dims;
     pub use super::udf::array_empty;
     pub use super::udf::array_length;
@@ -81,19 +84,17 @@ pub mod expr_fn {
     pub use super::udf::array_resize;
     pub use super::udf::array_reverse;
     pub use super::udf::array_sort;
-    pub use super::udf::array_to_string;
     pub use super::udf::cardinality;
     pub use super::udf::flatten;
     pub use super::udf::gen_series;
     pub use super::udf::range;
-    pub use super::udf::string_to_array;
 }
 
 /// Registers all enabled packages with a [`FunctionRegistry`]
 pub fn register_all(registry: &mut dyn FunctionRegistry) -> Result<()> {
     let functions: Vec<Arc<ScalarUDF>> = vec![
-        udf::array_to_string_udf(),
-        udf::string_to_array_udf(),
+        string::array_to_string_udf(),
+        string::string_to_array_udf(),
         udf::range_udf(),
         udf::gen_series_udf(),
         udf::array_dims_udf(),
