+++ datafusion/functions-array/src/utils.rs
@@ -214,6 +214,18 @@ pub(crate) fn compare_element_to_list(
     Ok(res)
 }
 
+macro_rules! downcast_arg {
+    ($ARG:expr, $ARRAY_TYPE:ident) => {{
+        $ARG.as_any().downcast_ref::<$ARRAY_TYPE>().ok_or_else(|| {
+            DataFusionError::Internal(format!(
+                "could not cast to {}",
+                type_name::<$ARRAY_TYPE>()
+            ))
+        })?
+    }};
+}
+pub(crate) use downcast_arg;
+
 #[cfg(test)]
 mod tests {
     use super::*;
